+++ rio/ChangeLog.md
@@ -5,6 +5,7 @@
 * Add `newLogFunc` function to create `LogFunc` records outside of a callback scope
 * Add `logGenericCallStack` to be able to register a log entry `CallStack` from a
   different thread
+* Add `getLogMinLevel` to gather what is the log min level from `LogOptions`
 * Allow dynamic reloading of `logMinLevel` and `logVerboseFormat` for the `LogOptions` record
 
 ## 0.1.2.0